Change semantic of P4_GET_CARD powers
The number of drawn card is encoded in the power:
- GET_CARD;
- GET_2_CARD;
- GET_3_CARD.
It is not affected by the value field of the power. This enables to tackle
a new power introduced by Xeno Invasion: consume for 2 vp and one card,
without resorting to the introduction of a GET_2_VP power. Furthermore,
the GET_XX_CARD powers were already in place, which minimizes the number of
required changes.