Check length of read strings to avoid buffer overflow.
Previously, received string length were determined by terminating '\0'
character. If none could be found, the reading buffer may be overflowed.
get_string now checks that the bound of the reading buffer are not exceeded

Additionally
- add macros to hold the length of reading buffers. This is done locally
  in ai_client.c, client.c and server.c because it is not uniform and
  the reason behind this non uniformity is not yet sorted out
- make the ai_client.c reading buff a static (internal) global variable.
  This allows for easier access by the other functions and is not an issue
  because only one buffer is processed at a time. In client.c, the reading
  buffer is duplicated and the copy is processed when the gtk loop has time
  available, which make turning the reading buffer into a static global
  variable useless.

Note: For the time being, the behaviour in case of detected buffer overflow
is compatible with the previous behaviour in the case of events that
shouldn't happen:
- the server kicks the clients which send such messages
- the client exits with an error message