Make the send/receive buffer of uniform length.
Server and clients shall be able to emit/receive message to/from each other
so the buffers for writing/reading messages shall be of the same size. We
stick to the minimal value present in the code, i.e., 1024 bytes.

Also use macro name, BUF_LEN instead of the value. Note: some char buffers
used while parsing messages are not properly message buffers. For
these buffer, we do not use the macro BUF_LEN