Add Rebel vs Imperium without Gathering Storm (RvIo)
Internally, it is encoded as a new (pseudo-)expansion.
Helper functions have been added to test properties of extensions.
- expansion_has_goals
- expansion_has_takeovers
- expansion_has_prestige
- expansion_has_start_world_choice

Previous expansion numbers not changed in order not to interfere with the
net mapping. Some code was added to set the order of presentation of the
expansions in the UI.

As for XI, server does not present RvIo games to clients older than 0.9.5